VERSION 01 finalisé - Ajouter des drones et les 3 status à la BD - Configurer le model HealthStatus dans HealthStatusEntityConfiguration et DroneDbcontext - créer un dbset pour recuperer les status des drones - recuperation des drones reussite dans la page Index.cs - Changer le type de HealthStatus ID : mettre decimale a la place de int -Lancement reussi, Tout est OK

+++ suivi-des-drones.Core.Models/HealthStatus.cs
@@ -18,7 +18,7 @@ namespace suivi_des_drones.Core.Models
         #endregion
 
         #region Properties
-        public int Id { get; init; }
+        public decimal Id { get; init; }
         public string Label { get; init; } = default!;
         #endregion
     }
